requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
debugpy==1.8.8
decorator==5.1.1
distro==1.9.0
elevenlabs==1.13.0
executing==2.1.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
-e git+https://github.com/rreben/LearnDutchWithGPT.git@eb171af1735473d43f7d2896d58bee00324f306c#egg=lesson_generator
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.54.4
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
requests==2.32.3
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
